Network.AWS.Pinpoint.CreateSegment

Description

Used to create or update a segment.

Creating a Request

createSegment

Arguments

:: Text
csApplicationId

-> WriteSegmentRequest
csWriteSegmentRequest

-> CreateSegment

Creates a value of CreateSegment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csApplicationId - Undocumented member.

csWriteSegmentRequest - Undocumented member.

data CreateSegment

See: createSegment smart constructor.

Request Lenses

csApplicationId :: Lens' CreateSegment Text

Undocumented member.

csWriteSegmentRequest :: Lens' CreateSegment WriteSegmentRequest

Undocumented member.

Destructuring the Response

createSegmentResponse

Arguments

:: Int
csrsResponseStatus

-> SegmentResponse
csrsSegmentResponse

-> CreateSegmentResponse

Creates a value of CreateSegmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csrsResponseStatus - -- | The response status code.

csrsSegmentResponse - Undocumented member.

data CreateSegmentResponse

See: createSegmentResponse smart constructor.

Response Lenses

csrsResponseStatus :: Lens' CreateSegmentResponse Int

- | The response status code.

csrsSegmentResponse :: Lens' CreateSegmentResponse SegmentResponse

Undocumented member.