Network.AWS.Pinpoint.DeleteAPNSSandboxChannel

Description

Delete an APNS sandbox channel

Creating a Request

deleteAPNSSandboxChannel

Arguments

:: Text
dascApplicationId

-> DeleteAPNSSandboxChannel

Creates a value of DeleteAPNSSandboxChannel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dascApplicationId - Undocumented member.

data DeleteAPNSSandboxChannel

See: deleteAPNSSandboxChannel smart constructor.

Request Lenses

dascApplicationId :: Lens' DeleteAPNSSandboxChannel Text

Undocumented member.

Destructuring the Response

deleteAPNSSandboxChannelResponse

Arguments

:: Int
dascrsResponseStatus

-> APNSSandboxChannelResponse
dascrsAPNSSandboxChannelResponse

-> DeleteAPNSSandboxChannelResponse

Creates a value of DeleteAPNSSandboxChannelResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dascrsResponseStatus - -- | The response status code.

dascrsAPNSSandboxChannelResponse - Undocumented member.

data DeleteAPNSSandboxChannelResponse

See: deleteAPNSSandboxChannelResponse smart constructor.

Response Lenses

dascrsResponseStatus :: Lens' DeleteAPNSSandboxChannelResponse Int

- | The response status code.

dascrsAPNSSandboxChannelResponse :: Lens' DeleteAPNSSandboxChannelResponse APNSSandboxChannelResponse

Undocumented member.