Network.AWS.Pinpoint.DeleteSegment

Description

Deletes a segment.

Creating a Request

deleteSegment

Arguments

:: Text
dsSegmentId

-> Text
dsApplicationId

-> DeleteSegment

Creates a value of DeleteSegment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsSegmentId - Undocumented member.

dsApplicationId - Undocumented member.

data DeleteSegment

See: deleteSegment smart constructor.

Request Lenses

dsSegmentId :: Lens' DeleteSegment Text

Undocumented member.

dsApplicationId :: Lens' DeleteSegment Text

Undocumented member.

Destructuring the Response

deleteSegmentResponse

Arguments

:: Int
dsrsResponseStatus

-> SegmentResponse
dsrsSegmentResponse

-> DeleteSegmentResponse

Creates a value of DeleteSegmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsResponseStatus - -- | The response status code.

dsrsSegmentResponse - Undocumented member.

data DeleteSegmentResponse

See: deleteSegmentResponse smart constructor.

Response Lenses

dsrsResponseStatus :: Lens' DeleteSegmentResponse Int

- | The response status code.

dsrsSegmentResponse :: Lens' DeleteSegmentResponse SegmentResponse

Undocumented member.