Network.AWS.Pinpoint.GetApplicationSettings

Description

Used to request the settings for an app.

Creating a Request

getApplicationSettings

Arguments

:: Text
gasApplicationId

-> GetApplicationSettings

Creates a value of GetApplicationSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gasApplicationId - Undocumented member.

data GetApplicationSettings

See: getApplicationSettings smart constructor.

Request Lenses

gasApplicationId :: Lens' GetApplicationSettings Text

Undocumented member.

Destructuring the Response

getApplicationSettingsResponse

Arguments

:: Int
gassrsResponseStatus

-> ApplicationSettingsResource
gassrsApplicationSettingsResource

-> GetApplicationSettingsResponse

Creates a value of GetApplicationSettingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gassrsResponseStatus - -- | The response status code.

gassrsApplicationSettingsResource - Undocumented member.

data GetApplicationSettingsResponse

See: getApplicationSettingsResponse smart constructor.

Response Lenses

gassrsResponseStatus :: Lens' GetApplicationSettingsResponse Int

- | The response status code.

gassrsApplicationSettingsResource :: Lens' GetApplicationSettingsResponse ApplicationSettingsResource

Undocumented member.