Network.AWS.Pinpoint.GetCampaigns

Description

Returns information about your campaigns.

Creating a Request

getCampaigns

Arguments

:: Text
gcsApplicationId

-> GetCampaigns

Creates a value of GetCampaigns with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcsToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gcsPageSize - The number of entries you want on each page in the response.

gcsApplicationId - Undocumented member.

data GetCampaigns

See: getCampaigns smart constructor.

Request Lenses

gcsToken :: Lens' GetCampaigns (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gcsPageSize :: Lens' GetCampaigns (Maybe Text)

The number of entries you want on each page in the response.

gcsApplicationId :: Lens' GetCampaigns Text

Undocumented member.

Destructuring the Response

getCampaignsResponse

Arguments

:: Int
gcsrsResponseStatus

-> CampaignsResponse
gcsrsCampaignsResponse

-> GetCampaignsResponse

Creates a value of GetCampaignsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcsrsResponseStatus - -- | The response status code.

gcsrsCampaignsResponse - Undocumented member.

data GetCampaignsResponse

See: getCampaignsResponse smart constructor.

Response Lenses

gcsrsResponseStatus :: Lens' GetCampaignsResponse Int

- | The response status code.

gcsrsCampaignsResponse :: Lens' GetCampaignsResponse CampaignsResponse

Undocumented member.