Network.AWS.Pinpoint.GetEmailChannel

Description

Get an email channel

Creating a Request

getEmailChannel

Arguments

:: Text
gecApplicationId

-> GetEmailChannel

Creates a value of GetEmailChannel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gecApplicationId - Undocumented member.

data GetEmailChannel

See: getEmailChannel smart constructor.

Request Lenses

gecApplicationId :: Lens' GetEmailChannel Text

Undocumented member.

Destructuring the Response

getEmailChannelResponse

Arguments

:: Int
gecrsResponseStatus

-> EmailChannelResponse
gecrsEmailChannelResponse

-> GetEmailChannelResponse

Creates a value of GetEmailChannelResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gecrsResponseStatus - -- | The response status code.

gecrsEmailChannelResponse - Undocumented member.

data GetEmailChannelResponse

See: getEmailChannelResponse smart constructor.

Response Lenses

gecrsResponseStatus :: Lens' GetEmailChannelResponse Int

- | The response status code.

gecrsEmailChannelResponse :: Lens' GetEmailChannelResponse EmailChannelResponse

Undocumented member.