Network.AWS.Pinpoint.GetImportJob

Description

Returns information about an import job.

Creating a Request

getImportJob

Arguments

:: Text
gijApplicationId

-> Text
gijJobId

-> GetImportJob

Creates a value of GetImportJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gijApplicationId - Undocumented member.

gijJobId - Undocumented member.

data GetImportJob

See: getImportJob smart constructor.

Request Lenses

gijApplicationId :: Lens' GetImportJob Text

Undocumented member.

gijJobId :: Lens' GetImportJob Text

Undocumented member.

Destructuring the Response

getImportJobResponse

Arguments

:: Int
gijrsResponseStatus

-> ImportJobResponse
gijrsImportJobResponse

-> GetImportJobResponse

Creates a value of GetImportJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gijrsResponseStatus - -- | The response status code.

gijrsImportJobResponse - Undocumented member.

data GetImportJobResponse

See: getImportJobResponse smart constructor.

Response Lenses

gijrsResponseStatus :: Lens' GetImportJobResponse Int

- | The response status code.

gijrsImportJobResponse :: Lens' GetImportJobResponse ImportJobResponse

Undocumented member.