Network.AWS.Pinpoint.GetImportJobs

Description

Returns information about your import jobs.

Creating a Request

getImportJobs

Arguments

:: Text
gijsApplicationId

-> GetImportJobs

Creates a value of GetImportJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gijsToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gijsPageSize - The number of entries you want on each page in the response.

gijsApplicationId - Undocumented member.

data GetImportJobs

See: getImportJobs smart constructor.

Request Lenses

gijsToken :: Lens' GetImportJobs (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gijsPageSize :: Lens' GetImportJobs (Maybe Text)

The number of entries you want on each page in the response.

gijsApplicationId :: Lens' GetImportJobs Text

Undocumented member.

Destructuring the Response

getImportJobsResponse

Arguments

:: Int
gijsrsResponseStatus

-> ImportJobsResponse
gijsrsImportJobsResponse

-> GetImportJobsResponse

Creates a value of GetImportJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gijsrsResponseStatus - -- | The response status code.

gijsrsImportJobsResponse - Undocumented member.

data GetImportJobsResponse

See: getImportJobsResponse smart constructor.

Response Lenses

gijsrsResponseStatus :: Lens' GetImportJobsResponse Int

- | The response status code.

gijsrsImportJobsResponse :: Lens' GetImportJobsResponse ImportJobsResponse

Undocumented member.