Network.AWS.Pinpoint.GetSegment

Description

Returns information about a segment.

Creating a Request

getSegment

Arguments

:: Text
gsSegmentId

-> Text
gsApplicationId

-> GetSegment

Creates a value of GetSegment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsSegmentId - Undocumented member.

gsApplicationId - Undocumented member.

data GetSegment

See: getSegment smart constructor.

Request Lenses

gsSegmentId :: Lens' GetSegment Text

Undocumented member.

gsApplicationId :: Lens' GetSegment Text

Undocumented member.

Destructuring the Response

getSegmentResponse

Arguments

:: Int
getrsResponseStatus

-> SegmentResponse
getrsSegmentResponse

-> GetSegmentResponse

Creates a value of GetSegmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

getrsResponseStatus - -- | The response status code.

getrsSegmentResponse - Undocumented member.

data GetSegmentResponse

See: getSegmentResponse smart constructor.

Response Lenses

getrsResponseStatus :: Lens' GetSegmentResponse Int

- | The response status code.

getrsSegmentResponse :: Lens' GetSegmentResponse SegmentResponse

Undocumented member.