Network.AWS.Pinpoint.GetSegmentVersion

Description

Returns information about a segment version.

Creating a Request

getSegmentVersion

Arguments

:: Text
gSegmentId

-> Text
gVersion

-> Text
gApplicationId

-> GetSegmentVersion

Creates a value of GetSegmentVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gSegmentId - Undocumented member.

gVersion - Undocumented member.

gApplicationId - Undocumented member.

data GetSegmentVersion

See: getSegmentVersion smart constructor.

Request Lenses

gSegmentId :: Lens' GetSegmentVersion Text

Undocumented member.

gVersion :: Lens' GetSegmentVersion Text

Undocumented member.

gApplicationId :: Lens' GetSegmentVersion Text

Undocumented member.

Destructuring the Response

getSegmentVersionResponse

Arguments

:: Int
gsvrsResponseStatus

-> SegmentResponse
gsvrsSegmentResponse

-> GetSegmentVersionResponse

Creates a value of GetSegmentVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsvrsResponseStatus - -- | The response status code.

gsvrsSegmentResponse - Undocumented member.

data GetSegmentVersionResponse

See: getSegmentVersionResponse smart constructor.

Response Lenses

gsvrsResponseStatus :: Lens' GetSegmentVersionResponse Int

- | The response status code.

gsvrsSegmentResponse :: Lens' GetSegmentVersionResponse SegmentResponse

Undocumented member.