Network.AWS.Pinpoint.GetSegmentVersions

Description

Returns information about your segment versions.

Creating a Request

getSegmentVersions

Arguments

:: Text
gsvSegmentId

-> Text
gsvApplicationId

-> GetSegmentVersions

Creates a value of GetSegmentVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsvToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gsvPageSize - The number of entries you want on each page in the response.

gsvSegmentId - Undocumented member.

gsvApplicationId - Undocumented member.

data GetSegmentVersions

See: getSegmentVersions smart constructor.

Request Lenses

gsvToken :: Lens' GetSegmentVersions (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gsvPageSize :: Lens' GetSegmentVersions (Maybe Text)

The number of entries you want on each page in the response.

gsvSegmentId :: Lens' GetSegmentVersions Text

Undocumented member.

gsvApplicationId :: Lens' GetSegmentVersions Text

Undocumented member.

Destructuring the Response

getSegmentVersionsResponse

Arguments

:: Int
grsResponseStatus

-> SegmentsResponse
grsSegmentsResponse

-> GetSegmentVersionsResponse

Creates a value of GetSegmentVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grsResponseStatus - -- | The response status code.

grsSegmentsResponse - Undocumented member.

data GetSegmentVersionsResponse

See: getSegmentVersionsResponse smart constructor.

Response Lenses

grsResponseStatus :: Lens' GetSegmentVersionsResponse Int

- | The response status code.

grsSegmentsResponse :: Lens' GetSegmentVersionsResponse SegmentsResponse

Undocumented member.