Network.AWS.Pinpoint.GetSegments

Description

Used to get information about your segments.

Creating a Request

getSegments

Arguments

:: Text
gssApplicationId

-> GetSegments

Creates a value of GetSegments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gssToken - The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gssPageSize - The number of entries you want on each page in the response.

gssApplicationId - Undocumented member.

data GetSegments

See: getSegments smart constructor.

Request Lenses

gssToken :: Lens' GetSegments (Maybe Text)

The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.

gssPageSize :: Lens' GetSegments (Maybe Text)

The number of entries you want on each page in the response.

gssApplicationId :: Lens' GetSegments Text

Undocumented member.

Destructuring the Response

getSegmentsResponse

Arguments

:: Int
gsrsResponseStatus

-> SegmentsResponse
gsrsSegmentsResponse

-> GetSegmentsResponse

Creates a value of GetSegmentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsrsResponseStatus - -- | The response status code.

gsrsSegmentsResponse - Undocumented member.

data GetSegmentsResponse

See: getSegmentsResponse smart constructor.

Response Lenses

gsrsResponseStatus :: Lens' GetSegmentsResponse Int

- | The response status code.

gsrsSegmentsResponse :: Lens' GetSegmentsResponse SegmentsResponse

Undocumented member.