Network.AWS.Pinpoint.PutEventStream

Description

Use to create or update the event stream for an app.

Creating a Request

putEventStream

Arguments

:: Text
pesApplicationId

-> WriteEventStream
pesWriteEventStream

-> PutEventStream

Creates a value of PutEventStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pesApplicationId - ApplicationId

pesWriteEventStream - EventStream to write.

data PutEventStream

PutEventStream Request

See: putEventStream smart constructor.

Request Lenses

pesApplicationId :: Lens' PutEventStream Text

ApplicationId

pesWriteEventStream :: Lens' PutEventStream WriteEventStream

EventStream to write.

Destructuring the Response

putEventStreamResponse

Arguments

:: Int
pesrsResponseStatus

-> EventStream
pesrsEventStream

-> PutEventStreamResponse

Creates a value of PutEventStreamResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pesrsResponseStatus - -- | The response status code.

pesrsEventStream - Undocumented member.

data PutEventStreamResponse

See: putEventStreamResponse smart constructor.

Response Lenses

pesrsResponseStatus :: Lens' PutEventStreamResponse Int

- | The response status code.

pesrsEventStream :: Lens' PutEventStreamResponse EventStream

Undocumented member.