Network.AWS.Pinpoint.SendMessages

Description

Send a batch of messages

Creating a Request

sendMessages

Arguments

:: Text
smApplicationId

-> MessageRequest
smMessageRequest

-> SendMessages

Creates a value of SendMessages with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smApplicationId - Undocumented member.

smMessageRequest - Undocumented member.

data SendMessages

See: sendMessages smart constructor.

Request Lenses

smApplicationId :: Lens' SendMessages Text

Undocumented member.

smMessageRequest :: Lens' SendMessages MessageRequest

Undocumented member.

Destructuring the Response

sendMessagesResponse

Arguments

:: Int
smrsResponseStatus

-> MessageResponse
smrsMessageResponse

-> SendMessagesResponse

Creates a value of SendMessagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smrsResponseStatus - -- | The response status code.

smrsMessageResponse - Undocumented member.

data SendMessagesResponse

See: sendMessagesResponse smart constructor.

Response Lenses

smrsResponseStatus :: Lens' SendMessagesResponse Int

- | The response status code.

smrsMessageResponse :: Lens' SendMessagesResponse MessageResponse

Undocumented member.