Network.AWS.Pinpoint.UpdateAPNSSandboxChannel

Description

Update an APNS sandbox channel

Creating a Request

updateAPNSSandboxChannel

Arguments

:: Text
uascApplicationId

-> APNSSandboxChannelRequest
uascAPNSSandboxChannelRequest

-> UpdateAPNSSandboxChannel

Creates a value of UpdateAPNSSandboxChannel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uascApplicationId - Undocumented member.

uascAPNSSandboxChannelRequest - Undocumented member.

data UpdateAPNSSandboxChannel

See: updateAPNSSandboxChannel smart constructor.

Request Lenses

uascApplicationId :: Lens' UpdateAPNSSandboxChannel Text

Undocumented member.

uascAPNSSandboxChannelRequest :: Lens' UpdateAPNSSandboxChannel APNSSandboxChannelRequest

Undocumented member.

Destructuring the Response

updateAPNSSandboxChannelResponse

Arguments

:: Int
uascrsResponseStatus

-> APNSSandboxChannelResponse
uascrsAPNSSandboxChannelResponse

-> UpdateAPNSSandboxChannelResponse

Creates a value of UpdateAPNSSandboxChannelResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uascrsResponseStatus - -- | The response status code.

uascrsAPNSSandboxChannelResponse - Undocumented member.

data UpdateAPNSSandboxChannelResponse

See: updateAPNSSandboxChannelResponse smart constructor.

Response Lenses

uascrsResponseStatus :: Lens' UpdateAPNSSandboxChannelResponse Int

- | The response status code.

uascrsAPNSSandboxChannelResponse :: Lens' UpdateAPNSSandboxChannelResponse APNSSandboxChannelResponse

Undocumented member.