Network.AWS.Pinpoint.UpdateEndpoint

Description

Use to update an endpoint.

Creating a Request

updateEndpoint

Arguments

:: Text
ueApplicationId

-> Text
ueEndpointId

-> EndpointRequest
ueEndpointRequest

-> UpdateEndpoint

Creates a value of UpdateEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ueApplicationId - Undocumented member.

ueEndpointId - Undocumented member.

ueEndpointRequest - Undocumented member.

data UpdateEndpoint

See: updateEndpoint smart constructor.

Request Lenses

ueApplicationId :: Lens' UpdateEndpoint Text

Undocumented member.

ueEndpointId :: Lens' UpdateEndpoint Text

Undocumented member.

ueEndpointRequest :: Lens' UpdateEndpoint EndpointRequest

Undocumented member.

Destructuring the Response

updateEndpointResponse

Arguments

:: Int
uersResponseStatus

-> MessageBody
uersMessageBody

-> UpdateEndpointResponse

Creates a value of UpdateEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uersResponseStatus - -- | The response status code.

uersMessageBody - Undocumented member.

data UpdateEndpointResponse

See: updateEndpointResponse smart constructor.

Response Lenses

uersResponseStatus :: Lens' UpdateEndpointResponse Int

- | The response status code.

uersMessageBody :: Lens' UpdateEndpointResponse MessageBody

Undocumented member.