Network.AWS.Pinpoint.UpdateEndpointsBatch

Description

Use to update a batch of endpoints.

Creating a Request

updateEndpointsBatch

Arguments

:: Text
uebApplicationId

-> EndpointBatchRequest
uebEndpointBatchRequest

-> UpdateEndpointsBatch

Creates a value of UpdateEndpointsBatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uebApplicationId - Undocumented member.

uebEndpointBatchRequest - Undocumented member.

data UpdateEndpointsBatch

See: updateEndpointsBatch smart constructor.

Request Lenses

uebApplicationId :: Lens' UpdateEndpointsBatch Text

Undocumented member.

uebEndpointBatchRequest :: Lens' UpdateEndpointsBatch EndpointBatchRequest

Undocumented member.

Destructuring the Response

updateEndpointsBatchResponse

Arguments

:: Int
uebrsResponseStatus

-> MessageBody
uebrsMessageBody

-> UpdateEndpointsBatchResponse

Creates a value of UpdateEndpointsBatchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uebrsResponseStatus - -- | The response status code.

uebrsMessageBody - Undocumented member.

data UpdateEndpointsBatchResponse

See: updateEndpointsBatchResponse smart constructor.

Response Lenses

uebrsResponseStatus :: Lens' UpdateEndpointsBatchResponse Int

- | The response status code.

uebrsMessageBody :: Lens' UpdateEndpointsBatchResponse MessageBody

Undocumented member.